+++ packages/liveblocks-client/src/LiveMap.test.ts
@@ -7,11 +7,11 @@ import {
   prepareStorageTest,
   reconnect,
 } from "../test/utils";
-import type { SerializedCrdtWithId } from "./live";
-import { CrdtType, OpCode, WebsocketCloseCodes } from "./live";
 import { LiveList } from "./LiveList";
 import { LiveMap } from "./LiveMap";
 import { LiveObject } from "./LiveObject";
+import type { SerializedCrdtWithId } from "./types";
+import { CrdtType, OpCode, WebsocketCloseCodes } from "./types";
 
 describe("LiveMap", () => {
   describe("not attached", () => {
