+++ packages/liveblocks-client/src/LiveMap.ts
@@ -4,12 +4,13 @@ import { errorIf } from "./deprecation";
 import type {
   CreateMapOp,
   CreateOp,
+  LiveMapUpdates,
+  Lson,
   Op,
   SerializedCrdt,
   SerializedCrdtWithId,
-} from "./live";
-import { CrdtType, OpCode } from "./live";
-import type { LiveMapUpdates, Lson } from "./types";
+} from "./types";
+import { CrdtType, OpCode } from "./types";
 import {
   creationOpToLiveStructure,
   deserialize,
