--- packages/liveblocks-client/src/live.ts
@@ -1,53 +1,0 @@
-export type {
-  BroadcastEventClientMsg,
-  ClientMsg,
-  FetchStorageClientMsg,
-  UpdatePresenceClientMsg,
-  UpdateStorageClientMsg,
-} from "./types/ClientMsg";
-export { ClientMsgCode } from "./types/ClientMsg";
-export type {
-  CreateListOp,
-  CreateMapOp,
-  CreateObjectOp,
-  CreateOp,
-  CreateRegisterOp,
-  DeleteCrdtOp,
-  DeleteObjectKeyOp,
-  Op,
-  SetParentKeyOp,
-  UpdateObjectOp,
-} from "./types/Op";
-export { OpCode } from "./types/Op";
-export type {
-  SerializedCrdt,
-  SerializedList,
-  SerializedMap,
-  SerializedObject,
-  SerializedRegister,
-} from "./types/SerializedCrdt";
-export { CrdtType } from "./types/SerializedCrdt";
-export type {
-  BroadcastedEventServerMsg,
-  InitialDocumentStateServerMsg,
-  RoomStateServerMsg,
-  SerializedCrdtWithId,
-  ServerMsg,
-  UpdatePresenceServerMsg,
-  UpdateStorageServerMsg,
-  UserJoinServerMsg,
-  UserLeftServerMsg,
-} from "./types/ServerMsg";
-export { ServerMsgCode } from "./types/ServerMsg";
-
-export enum WebsocketCloseCodes {
-  CLOSE_ABNORMAL = 1006,
-
-  INVALID_MESSAGE_FORMAT = 4000,
-  NOT_ALLOWED = 4001,
-  MAX_NUMBER_OF_MESSAGES_PER_SECONDS = 4002,
-  MAX_NUMBER_OF_CONCURRENT_CONNECTIONS = 4003,
-  MAX_NUMBER_OF_MESSAGES_PER_DAY_PER_APP = 4004,
-  MAX_NUMBER_OF_CONCURRENT_CONNECTIONS_PER_ROOM = 4005,
-  CLOSE_WITHOUT_RETRY = 4999,
-}
