+++ packages/liveblocks-client/src/room.ts
@@ -1,24 +1,5 @@
 import type { ApplyResult } from "./AbstractCrdt";
 import { AbstractCrdt } from "./AbstractCrdt";
-import type {
-  BroadcastedEventServerMsg,
-  ClientMsg,
-  InitialDocumentStateServerMsg,
-  Op,
-  RoomStateServerMsg,
-  SerializedCrdt,
-  SerializedCrdtWithId,
-  ServerMsg,
-  UpdatePresenceServerMsg,
-  UserJoinServerMsg,
-  UserLeftServerMsg,
-} from "./live";
-import {
-  ClientMsgCode,
-  OpCode,
-  ServerMsgCode,
-  WebsocketCloseCodes,
-} from "./live";
 import { LiveList } from "./LiveList";
 import type { LiveMap } from "./LiveMap";
 import { LiveObject } from "./LiveObject";
@@ -26,17 +7,21 @@ import type {
   Authentication,
   AuthenticationToken,
   AuthorizeResponse,
+  BroadcastedEventServerMsg,
   BroadcastOptions,
+  ClientMsg,
   Connection,
   ConnectionCallback,
   ConnectionState,
   ErrorCallback,
   EventCallback,
+  InitialDocumentStateServerMsg,
   Json,
   JsonObject,
   Lson,
   LsonObject,
   MyPresenceCallback,
+  Op,
   Others,
   OthersEvent,
   OthersEventCallback,
@@ -45,9 +30,22 @@ import type {
   RoomEventCallbackMap,
   RoomEventName,
   RoomInitializers,
+  RoomStateServerMsg,
+  SerializedCrdt,
+  SerializedCrdtWithId,
+  ServerMsg,
   StorageCallback,
   StorageUpdate,
+  UpdatePresenceServerMsg,
   User,
+  UserJoinServerMsg,
+  UserLeftServerMsg,
+} from "./types";
+import {
+  ClientMsgCode,
+  OpCode,
+  ServerMsgCode,
+  WebsocketCloseCodes,
 } from "./types";
 import { isJsonArray, isJsonObject, parseJson } from "./types/Json";
 import {
