+++ packages/liveblocks-client/src/types/index.ts
@@ -635,5 +635,58 @@ export type Room = {
   batch: (fn: () => void) => void;
 };
 
+export enum WebsocketCloseCodes {
+  CLOSE_ABNORMAL = 1006,
+
+  INVALID_MESSAGE_FORMAT = 4000,
+  NOT_ALLOWED = 4001,
+  MAX_NUMBER_OF_MESSAGES_PER_SECONDS = 4002,
+  MAX_NUMBER_OF_CONCURRENT_CONNECTIONS = 4003,
+  MAX_NUMBER_OF_MESSAGES_PER_DAY_PER_APP = 4004,
+  MAX_NUMBER_OF_CONCURRENT_CONNECTIONS_PER_ROOM = 4005,
+  CLOSE_WITHOUT_RETRY = 4999,
+}
+
+export type {
+  BroadcastEventClientMsg,
+  ClientMsg,
+  FetchStorageClientMsg,
+  UpdatePresenceClientMsg,
+  UpdateStorageClientMsg,
+} from "./ClientMsg";
+export { ClientMsgCode } from "./ClientMsg";
 export type { Json, JsonObject } from "./Json";
 export type { Lson, LsonObject, ToJson } from "./Lson";
+export type {
+  CreateListOp,
+  CreateMapOp,
+  CreateObjectOp,
+  CreateOp,
+  CreateRegisterOp,
+  DeleteCrdtOp,
+  DeleteObjectKeyOp,
+  Op,
+  SetParentKeyOp,
+  UpdateObjectOp,
+} from "./Op";
+export { OpCode } from "./Op";
+export type {
+  SerializedCrdt,
+  SerializedList,
+  SerializedMap,
+  SerializedObject,
+  SerializedRegister,
+} from "./SerializedCrdt";
+export { CrdtType } from "./SerializedCrdt";
+export type {
+  BroadcastedEventServerMsg,
+  InitialDocumentStateServerMsg,
+  RoomStateServerMsg,
+  SerializedCrdtWithId,
+  ServerMsg,
+  UpdatePresenceServerMsg,
+  UpdateStorageServerMsg,
+  UserJoinServerMsg,
+  UserLeftServerMsg,
+} from "./ServerMsg";
+export { ServerMsgCode } from "./ServerMsg";
