+++ packages/liveblocks-client/src/utils.ts
@@ -1,27 +1,25 @@
 import type { AbstractCrdt, Doc } from "./AbstractCrdt";
-import type {
-  CreateOp,
-  Op,
-  SerializedCrdt,
-  SerializedCrdtWithId,
-  SerializedList,
-  SerializedMap,
-  SerializedObject,
-} from "./live";
-import { CrdtType, OpCode } from "./live";
 import { LiveList } from "./LiveList";
 import { LiveMap } from "./LiveMap";
 import { LiveObject } from "./LiveObject";
 import { LiveRegister } from "./LiveRegister";
 import type {
+  CreateOp,
   Json,
   LiveListUpdates,
   LiveMapUpdates,
   LiveObjectUpdates,
   Lson,
   LsonObject,
+  Op,
+  SerializedCrdt,
+  SerializedCrdtWithId,
+  SerializedList,
+  SerializedMap,
+  SerializedObject,
   StorageUpdate,
 } from "./types";
+import { CrdtType, OpCode } from "./types";
 import { isJsonObject, parseJson } from "./types/Json";
 
 export function remove<T>(array: T[], item: T) {
