+++ packages/liveblocks-client/test/utils.ts
@@ -1,22 +1,20 @@
 import type { AbstractCrdt } from "../src/AbstractCrdt";
 import { lsonToJson, patchImmutableObject } from "../src/immutable";
-import type {
-  ClientMsg,
-  Op,
-  SerializedCrdtWithId,
-  ServerMsg,
-} from "../src/live";
-import { ClientMsgCode, CrdtType, ServerMsgCode } from "../src/live";
 import { makePosition } from "../src/position";
 import type { Effects, Machine } from "../src/room";
 import { defaultState, makeStateMachine } from "../src/room";
 import type {
   Authentication,
+  ClientMsg,
   Json,
   JsonObject,
   LsonObject,
+  Op,
+  SerializedCrdtWithId,
+  ServerMsg,
   ToJson,
 } from "../src/types";
+import { ClientMsgCode, CrdtType, ServerMsgCode } from "../src/types";
 import { remove } from "../src/utils";
 
 /**
